+++ src/msgbuf.h
@@ -1023,6 +1023,27 @@ pldm_msgbuf_span_required(struct pldm_msgbuf *ctx, size_t required,
 }
 
 __attribute__((always_inline)) static inline int
+pldm_msgbuf_span_string_ascii(struct pldm_msgbuf *ctx, void **cursor)
+{
+	assert(ctx);
+
+	if (!ctx->cursor || !cursor || *cursor) {
+		return pldm_msgbuf_status(ctx, EINVAL);
+	}
+
+	ctx->remaining -= (intmax_t)strlen((char *)ctx->cursor) + 1;
+	assert(ctx->remaining >= 0);
+	if (ctx->remaining < 0) {
+		return pldm_msgbuf_status(ctx, EINVAL);
+	}
+
+	*cursor = (char *)ctx->cursor;
+	ctx->cursor += strlen((char *)ctx->cursor) + 1;
+
+	return 0;
+}
+
+__attribute__((always_inline)) static inline int
 pldm_msgbuf_span_remaining(struct pldm_msgbuf *ctx, void **cursor, size_t *len)
 {
 	assert(ctx);